 requirements.txt
google_speech
SpeechRecognition
pydub
pocketsphinx
pinecone-client
datasets
transformers